week7/covid19-research-explorer/requirements.txt
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
streamlit==1.28.1
plotly==5.17.0
wordcloud==1.9.2
scikit-learn==1.3.0
scipy==1.11.2
python-dotenv==1.0.0
pyyaml==6.0.1